feat(proxy): add a proxy list component and retrieve created proxy entity at creation
