+++ frontend/src/app/engine/engine.component.spec.ts
@@ -2,16 +2,13 @@ import { TestBed } from '@angular/core/testing';
 
 import { EngineComponent } from './engine.component';
 import { ComponentTester, createMock } from 'ngx-speculoos';
-import { provideTestingI18n } from '../../i18n/mock-i18n';
-import { provideRouter } from '@angular/router';
-import { provideHttpClient } from '@angular/common/http';
-import { EngineSettingsCommandDTO, EngineSettingsDTO } from '../model/engine.model';
-import { EngineService } from '../services/engine.service';
-import { ProxyService } from '../services/proxy.service';
-import { ProxyDTO } from '../model/proxy.model';
+import { EngineSettingsDTO } from '../model/engine.model';
 import { of } from 'rxjs';
-import { NotificationService } from '../components/shared/notification.service';
-import { SaveButtonComponent } from '../components/shared/save-button/save-button.component';
+import { EngineService } from '../services/engine.service';
+import { provideTestingI18n } from '../../i18n/mock-i18n';
+import { ProxyListComponent } from './proxy-list/proxy-list.component';
+import { provideHttpClient } from '@angular/common/http';
+import { provideRouter } from '@angular/router';
 
 class EngineComponentTester extends ComponentTester<EngineComponent> {
   constructor() {
@@ -23,103 +20,49 @@ class EngineComponentTester extends ComponentTester<EngineComponent> {
   }
 
   get name() {
-    return this.input('#name')!;
+    return this.element('#name')!;
   }
 
   get port() {
-    return this.input('#port')!;
+    return this.element('#port')!;
   }
 
   get consoleLevel() {
-    return this.select('#console-level')!;
+    return this.element('#console-level')!;
   }
 
   get fileLevel() {
-    return this.select('#file-level')!;
-  }
-
-  get fileMaxFileSize() {
-    return this.input('#file-max-file-size');
-  }
-
-  get fileNumberOfFiles() {
-    return this.input('#file-number-of-files');
+    return this.element('#file-level')!;
   }
 
   get databaseLevel() {
-    return this.select('#database-level')!;
-  }
-
-  get databaseMaxNumberOfLogs() {
-    return this.input('#database-max-number-of-logs');
+    return this.element('#database-level')!;
   }
 
   get lokiLevel() {
-    return this.select('#loki-level')!;
-  }
-
-  get lokiInterval() {
-    return this.input('#loki-interval');
-  }
-
-  get lokiAddress() {
-    return this.input('#loki-address');
-  }
-
-  get lokiProxy() {
-    return this.select('#loki-proxy');
-  }
-
-  get lokiTokenAddress() {
-    return this.input('#loki-token-address');
-  }
-
-  get lokiUsername() {
-    return this.input('#loki-username');
-  }
-
-  get lokiPassword() {
-    return this.input('#loki-password');
+    return this.element('#loki-level')!;
   }
 
   get healthSignalLoggingEnabled() {
-    return this.input('#health-signal-logging-enabled')!;
-  }
-
-  get healthSignalLoggingInterval() {
-    return this.input('#health-signal-logging-interval');
+    return this.element('#health-logging')!;
   }
 
   get healthSignalHttpEnabled() {
-    return this.input('#health-signal-http-enabled')!;
+    return this.element('#health-http')!;
   }
 
-  get healthSignalHttpVerbose() {
-    return this.input('#health-signal-http-verbose');
+  get editButton() {
+    return this.button('#edit-link')!;
   }
 
-  get healthSignalHttpInterval() {
-    return this.input('#health-signal-http-interval');
-  }
-
-  get healthSignalHttpAddress() {
-    return this.input('#health-signal-http-address');
-  }
-
-  get healthSignalHttpProxy() {
-    return this.select('#health-signal-http-proxy');
-  }
-
-  get submitButton() {
-    return this.button('#save-button')!;
+  get proxyList() {
+    return this.element(ProxyListComponent);
   }
 }
 
 describe('EngineComponent', () => {
   let tester: EngineComponentTester;
   let engineService: jasmine.SpyObj<EngineService>;
-  let proxyService: jasmine.SpyObj<ProxyService>;
-  let notificationService: jasmine.SpyObj<NotificationService>;
 
   const engineSettings: EngineSettingsDTO = {
     id: 'id',
@@ -130,196 +73,51 @@ describe('EngineComponent', () => {
         level: 'silent'
       },
       file: {
-        level: 'trace',
-        numberOfFiles: 5,
-        maxFileSize: 10
+        level: 'trace'
       },
       database: {
-        level: 'silent',
-        maxNumberOfLogs: 100_000
+        level: 'silent'
       },
       loki: {
-        level: 'error',
-        interval: 60,
-        address: 'http://loki.oibus.com',
-        tokenAddress: 'http://token-address.oibus.com',
-        username: 'oibus',
-        password: 'pass',
-        proxyId: null
+        level: 'error'
       }
     },
     healthSignal: {
       logging: {
-        enabled: true,
-        interval: 60
+        enabled: true
       },
       http: {
-        enabled: true,
-        interval: 60,
-        verbose: false,
-        address: 'http://health-signal.oibus.com',
-        proxyId: null,
-        authentication: {
-          type: 'basic',
-          key: 'oibus',
-          secret: 'pass'
-        }
+        enabled: true
       }
     }
-  };
-  const proxies: Array<ProxyDTO> = [
-    {
-      id: 'id1',
-      name: 'proxy1',
-      description: 'My Proxy 1',
-      address: 'http://localhost',
-      username: 'user',
-      password: 'pass'
-    },
-    {
-      id: 'id2',
-      name: 'proxy2',
-      description: 'My Proxy 2',
-      address: 'http://localhost',
-      username: 'user',
-      password: 'pass'
-    }
-  ];
+  } as EngineSettingsDTO;
 
   beforeEach(() => {
     engineService = createMock(EngineService);
-    proxyService = createMock(ProxyService);
-    notificationService = createMock(NotificationService);
 
     TestBed.configureTestingModule({
-      imports: [EngineComponent, SaveButtonComponent],
-      providers: [
-        provideHttpClient(),
-        provideRouter([]),
-        provideTestingI18n(),
-        { provide: EngineService, useValue: engineService },
-        { provide: ProxyService, useValue: proxyService },
-        { provide: NotificationService, useValue: notificationService }
-      ]
-    });
+      imports: [EngineComponent, ProxyListComponent],
+      providers: [provideTestingI18n(), provideRouter([]), provideHttpClient(), { provide: EngineService, useValue: engineService }]
+    }).compileComponents();
 
     engineService.getEngineSettings.and.returnValue(of(engineSettings));
-
-    engineService.updateEngineSettings.and.returnValue(of(undefined));
-    proxyService.getProxies.and.returnValue(of(proxies));
 
     tester = new EngineComponentTester();
     tester.detectChanges();
   });
 
-  it('should display title and filled form', () => {
+  it('should display engine settings', () => {
     expect(tester.title).toContainText('Engine');
-    expect(tester.name).toHaveValue(engineSettings.name);
-    expect(tester.port).toHaveValue(engineSettings.port.toString());
-    expect(tester.consoleLevel).toHaveSelectedLabel('Silent');
-    expect(tester.fileLevel).toHaveSelectedLabel('Trace');
-    expect(tester.fileMaxFileSize).toHaveValue(engineSettings.logParameters.file.maxFileSize.toString());
-    expect(tester.fileNumberOfFiles).toHaveValue(engineSettings.logParameters.file.numberOfFiles.toString());
-    expect(tester.databaseLevel).toHaveSelectedLabel('Silent');
-    expect(tester.databaseMaxNumberOfLogs).toBeNull();
-    expect(tester.lokiLevel).toHaveSelectedLabel('Error');
-    expect(tester.lokiInterval).toHaveValue(engineSettings.logParameters.loki.interval.toString());
-    expect(tester.lokiAddress).toHaveValue(engineSettings.logParameters.loki.address);
-    expect(tester.lokiTokenAddress).toHaveValue(engineSettings.logParameters.loki.tokenAddress);
-    expect(tester.lokiAddress).toHaveValue(engineSettings.logParameters.loki.address);
-    expect(tester.lokiUsername).toHaveValue(engineSettings.logParameters.loki.username);
-    expect(tester.lokiPassword).toHaveValue(engineSettings.logParameters.loki.password);
-    expect(tester.lokiProxy).toHaveSelectedLabel('');
-    expect(tester.healthSignalLoggingEnabled).toBeChecked();
-    expect(tester.healthSignalLoggingInterval).toHaveValue(engineSettings.healthSignal.logging.interval.toString());
-    expect(tester.healthSignalHttpEnabled).toBeChecked();
-    expect(tester.healthSignalHttpVerbose).not.toBeChecked();
-    expect(tester.healthSignalHttpInterval).toHaveValue(engineSettings.healthSignal.http.interval.toString());
-    expect(tester.healthSignalHttpAddress).toHaveValue(engineSettings.healthSignal.http.address);
-    expect(tester.healthSignalHttpProxy).toHaveSelectedLabel('');
-  });
+    expect(tester.name).toContainText('OIBus Test');
+    expect(tester.port).toContainText('Port: 2223');
+    expect(tester.consoleLevel).toContainText('Console: Silent');
+    expect(tester.fileLevel).toContainText('File: Trace');
+    expect(tester.databaseLevel).toContainText('Database: Silent');
+    expect(tester.lokiLevel).toContainText('Loki: Error');
+    expect(tester.healthSignalLoggingEnabled).toContainText('Logger: Enabled');
+    expect(tester.healthSignalHttpEnabled).toContainText('HTTP: Enabled');
+    expect(tester.editButton).toContainText('Edit settings');
 
-  it('should update engine settings', () => {
-    tester.name.fillWith('OIBus Dev');
-
-    tester.consoleLevel.selectLabel('Error');
-    tester.fileNumberOfFiles!.fillWith('10');
-    tester.lokiLevel.selectLabel('Silent');
-    tester.healthSignalLoggingEnabled.uncheck();
-    tester.healthSignalHttpVerbose!.check();
-
-    expect(tester.name).toHaveValue('OIBus Dev');
-    expect(tester.consoleLevel).toHaveSelectedLabel('Error');
-    expect(tester.fileNumberOfFiles).toHaveValue('10');
-    expect(tester.lokiLevel).toHaveSelectedLabel('Silent');
-    expect(tester.lokiInterval).toBeNull();
-    expect(tester.lokiAddress).toBeNull();
-    expect(tester.lokiTokenAddress).toBeNull();
-    expect(tester.lokiAddress).toBeNull();
-    expect(tester.lokiUsername).toBeNull();
-    expect(tester.lokiPassword).toBeNull();
-    expect(tester.lokiProxy).toBeNull();
-    expect(tester.healthSignalLoggingEnabled).not.toBeChecked();
-    expect(tester.healthSignalLoggingInterval).toBeNull();
-    expect(tester.healthSignalHttpEnabled).toBeChecked();
-    expect(tester.healthSignalHttpVerbose).toBeChecked();
-
-    tester.submitButton.click();
-
-    const expectedSettings: EngineSettingsCommandDTO = JSON.parse(JSON.stringify(engineSettings));
-
-    expectedSettings.name = 'OIBus Dev';
-    expectedSettings.logParameters.console.level = 'error';
-    expectedSettings.logParameters.file.numberOfFiles = 10;
-    expectedSettings.logParameters.loki.level = 'silent';
-    expectedSettings.healthSignal.logging.enabled = false;
-    expectedSettings.healthSignal.http.verbose = true;
-
-    expect(engineService.updateEngineSettings).toHaveBeenCalledWith({
-      name: 'OIBus Dev',
-      port: 2223,
-      logParameters: {
-        console: {
-          level: 'error'
-        },
-        file: {
-          level: 'trace',
-          numberOfFiles: 10,
-          maxFileSize: 10
-        },
-        database: {
-          level: 'silent',
-          maxNumberOfLogs: 100_000
-        },
-        loki: {
-          level: 'silent',
-          interval: 60,
-          address: 'http://loki.oibus.com',
-          tokenAddress: 'http://token-address.oibus.com',
-          username: 'oibus',
-          password: 'pass',
-          proxyId: null
-        }
-      },
-      healthSignal: {
-        logging: {
-          enabled: false,
-          interval: 60
-        },
-        http: {
-          enabled: true,
-          interval: 60,
-          verbose: true,
-          address: 'http://health-signal.oibus.com',
-          proxyId: null,
-          authentication: {
-            type: 'basic',
-            key: 'oibus',
-            secret: 'pass'
-          }
-        }
-      }
-    });
-    expect(notificationService.success).toHaveBeenCalledWith('engine.updated');
+    expect(tester.proxyList).toBeDefined();
   });
 });
