+++ frontend/src/app/engine/proxy-list/proxy-list.component.html
@@ -1,0 +1,37 @@
+<h2 translate="engine.proxy.list">
+  <button class="btn btn-primary ms-3" id="add-proxy" (click)="openCreationProxyModal()">
+    <span class="fa fa-plus me-2"></span>
+    <span translate="engine.proxy.add"></span>
+  </button>
+</h2>
+<!-- the proxy list -->
+<div *ngIf="proxies.length !== 0; else emptyProxies">
+  <table class="mt-2 table table-sm table-hover">
+    <thead class="table-light">
+      <tr>
+        <th translate="engine.proxy.name"></th>
+        <th translate="engine.proxy.description"></th>
+        <th translate="engine.proxy.address"></th>
+        <th style="width: 3%"></th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr *ngFor="let proxy of proxies">
+        <td>{{ proxy.name }}</td>
+        <td>{{ proxy.description }}</td>
+        <td>{{ proxy.address }}</td>
+        <td class="text-nowrap action-buttons">
+          <div class="pull-right">
+            <!-- Edit button -->
+            <button type="button" class="btn btn-link edit-proxy fa fa-pencil px-1 py-0" (click)="openEditProxyModal(proxy)"></button>
+            <!-- Delete button -->
+            <button type="button" class="btn btn-link proxy fa fa-trash px-1 py-0" (click)="deleteProxy(proxy)"></button>
+          </div>
+        </td>
+      </tr>
+    </tbody>
+  </table>
+</div>
+<ng-template #emptyProxies>
+  <div class="oi-details empty" id="no-proxy" translate="engine.proxy.none"></div>
+</ng-template>
